Deduplicate entities derived from GroupEntity
